fixed function calls with default args, added run_all repros

+++ repro/run.sh
@@ -1,2 +1,8 @@
-ela $1.ela && ./$1
-rm $1
+if ela $1.ela; then
+	./$1
+	run_status=$?
+	rm $1
+	exit $run_status
+else
+	exit 1
+fi
